requirements.txt
certifi==2020.6.20
chainer==7.7.0
cupy==7.8.0.post1
cycler==0.10.0
decorator==4.4.2
fastrlock==0.5
filelock==3.0.12
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.2.0
matplotlib==3.3.1
msgpack==1.0.2
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.42
Pillow==7.2.0
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.2
Shapely==1.7.1
six==1.15.0
tensorboardX==2.1
tifffile==2020.9.3
tqdm==4.51.0
typing-extensions==3.7.4.3
